% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawped2sex.R
\name{buildped.2sexes}
\alias{buildped.2sexes}
\title{Builds a two-sex pedigree}
\usage{
buildped.2sex(founders.m=50,founders.f=50,fert=2.0,death.rate=0.2,breed.prop=0.5,male.prop=0.5,n.tstep=10)
}
\arguments{
\item{founders.m}{numbers of male founders}

\item{founders.f}{number of female founders}

\item{fert}{female fertility rate per time interval}

\item{death.rate}{mortality rate per time interval}

\item{breed.prop}{proportion of breeding males}

\item{male.prop}{proportion of males (sex ratio)}

\item{n.tstep}{number of time step}
}
\value{
a pedigree data frame with 4 columns [ind id, dam id, sire id, sex] and as many rows as individuals
}
\description{
Builds a two-sex pedigree from the passed arguments
}
\details{

}
\examples{
\dontrun{

#build a pedigree
library(hierfstat)
test<-buildped.2sexes(founders.m=10,founders.f=10,fert=1,death.rate=0.5,breed.prop=0.2,n.tstep=5)
#range of kinship coefficients
range(mat2vec(grm2kinship(pedARM(test[,2],test[,3]))))
#range of inbreeding coefficients
range(diag(grm2kinship(pedARM(test[,2],test[,3]))))
#plot the pedigree
library(kinship2)
with(test,plot(pedigree(ind,sire,dam,sex)))
}
}
\references{

}
\author{
Jerome Goudet \email{jerome.goudet@unil.ch}
}
