% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawped.herma.R
\name{buildped.herma}
\alias{buildped.herma}
\title{Builds a hermaphrodite pedigree}
\usage{
buildped.herma(founders=20,fert=0.5,death.rate=0.5,self.prop=0.5,n.tstep=10)
}
\arguments{
\item{founders}{numbers of male founders}

\item{fert}{hermaphrodite fertility rate per time interval}

\item{death.rate}{mortality rate per time interval}

\item{self.prop}{proportion of selfing}

\item{n.tstep}{number of time step}
}
\value{
a pedigree data frame with 3 columns [ind id, parent1 id, parent2 id] and as many rows as individuals
}
\description{
Builds a hermaphrodite pedigree from the passed arguments
}
\details{

}
\examples{
\dontrun{

#build a pedigree
library(hierfstat)
test<-buildped.herma(founders=20,fert=0.5,death.rate=0.5,self.prop=0.5,n.tstep=5)
#range of kinship coefficients
range(mat2vec(grm2kinship(pedARM(test[,2],test[,3]))))
#range of inbreeding coefficients
range(diag(grm2kinship(pedARM(test[,2],test[,3]))))
}
}
\references{

}
\author{
Jerome Goudet \email{jerome.goudet@unil.ch}
}
