% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp.ci.binom.R
\name{comp.ci.binom}
\alias{comp.ci.binom}
\title{Compares Confidence intervals of the binomial distribution}
\usage{
comp.ci.binom(p=0.5,n=1000,nboot=1000,digits=4)
}
\arguments{
\item{p}{the binomial probabilty of success (frequency)}

\item{n}{How many draws}

\item{nboot}{number of bootstrap samples}

\item{digits}{how many digits to print out}
}
\value{
a list with components
\itemize{
\item p.hat
\item norm.ci Wald confidence interval
\item boot.ci
\item exact.ci
}
}
\description{
computes 95% confidence intervals of the probability of
success using the normal approximation, bootstrapping
and an exact confidence interval
}
\examples{
{
comp.ci.binom()
comp.ci.binom(n=10,p=0.1)
}

}
\seealso{
the corresponding \href{https://jgx65.shinyapps.io/BinomNorm/}{shiny app}
}
