% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drift.R
\name{drift}
\alias{drift}
\title{Simulates drift}
\usage{
drift (nind=100, ngen=100, p0=0.5, nrep=1000, PlotIt=FALSE)
}
\arguments{
\item{nind}{number of individuals}

\item{ngen}{number of generations}

\item{p0}{initial allel frequency}

\item{nrep}{number of replicates}

\item{PlotIt}{whether to plot the allelic trajectories}
}
\value{
a matrix with the frequencies of replicates in column for each generation in rows
}
\description{
Simulates drift
}
\examples{
\dontrun{
res<-drift(ngen=500)
par(mfrow=c(3,3))
for (i in c(1,2,5,10,20,50,100,200,500))
hist(res[i,],breaks=0:100/100,main=paste0('gen: ',i),xlab="p",ylab="")
}

}
