% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropAlongPed.R
\name{drop.along.ped}
\alias{drop.along.ped}
\title{drops alleles along pedigree}
\usage{
drop.along.ped(ped=ped,founders.genotypes=NULL,nloc=10000,maplength=20)
}
\arguments{
\item{ped}{a pedigree with at least three columns, the first for the individual id,
the second for the dam id and the third for the sire id.
dams and sires of founders are entered as NAs
founders/unknown parents must appear first in the pedigree}

\item{founders.genotypes}{(default NULL) if given, an array [nf*2,nloc] containing the (0 or 1)
of the nf founders at nloc loci. if only one parent is missing,
the missing parent should be considered as a founder}

\item{nloc}{number of equally spaced loci}

\item{maplength}{map length in Morgans}
}
\value{
a [ni,nl,2] array of genotypes
}
\description{
Given a pedigree, drops alleles along the pedigree using the given recombination map.
Could be seeded with genotypes for the founder individuals, otherwise, uses gene dropping.
}
\examples{
\dontrun{}
}
\author{
Jerome Goudet \email{jerome.goudet@unil.ch}
}
