% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFuniW.R
\name{get.funiu}
\alias{get.funiu}
\title{Estimates FUNI (FIII) as average of ratios}
\usage{
get.funiu(dos)
}
\arguments{
\item{dos}{a \code{\link[gaston]{bed.matrix}} object}
}
\value{
a list with components
\itemize{
\item nloc the number of typed loci for each individual
\item Funi average of ratios \eqn{F_{uni}} estimates
}
}
\description{
Estimates FUNI as average of ratios on a \code{\link[gaston]{bed.matrix}} object
}
\details{
This is the original definition of FIII or FUNI given in Yang et al. (2010, 2011)
}
\references{
{

\href{https://doi.org/10.1038/s41437-021-00471-4}{Zhang QS, Goudet J and Weir BS 2022}
 Rank-invariant Estimation of Inbreeding coefficients. Heredity

Original definitions of FUNI and FIII are found in:

     Yang J, Benyamin B, McEvoy BP, Gordon S, Henders AK, Nyholt DR,
Madden PA, Heath AC, Martin NG, Montgomery GW, Goddard ME,
Visscher PM. 2010.  Common SNPs explain a large proportion of the
heritability for human height.  Nat Genet. 42(7):565-9. Epub 2010
Jun 20.

Yang, J., Lee, S. H., Goddard, M. E. & Visscher, P. M.  GCTA: a
tool for genome-wide complex trait analysis.  American journal of
human genetics 88, 76-82 (2011).

}
}
\author{
Jerome Goudet \email{jerome.goudet@unil.ch}
}
