% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFuniW.R
\name{get.funiw}
\alias{get.funiw}
\title{Estimates FUNI as ratio of averages}
\usage{
get.funiw(dos)
}
\arguments{
\item{dos}{a \code{\link[gaston]{bed.matrix}} object}
}
\value{
a list with components
\itemize{
\item het the sum of loci heterozygosities
\item Funi ratio of averages \eqn{F_{uni}} estimates
}
}
\description{
Estimates FUNI as ratio of averages on a \code{\link[gaston]{bed.matrix}} object
}
\details{



}
\examples{
{\dontrun{}}

}
\references{
\href{https://doi.org/10.1038/s41437-021-00471-4}{Zhang QS, Goudet J and Weir BS 2022}
 Rank-invariant Estimation of Inbreeding coefficients. Heredity
}
\author{
Jerome Goudet \email{jerome.goudet@unil.ch}
}
