% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropAlongPed.R
\name{jaccard}
\alias{jaccard}
\title{Calculates the 9 Jaccard's Delta coefficients}
\usage{
jaccard(dat)
}
\arguments{
\item{dat}{an nixnlx2 array of genotypes ( ni individuals, nl loci) generated with \code{\link{drop.along.ped}}}
}
\value{
a list of 9 matrices, the lower half of which contain
the corresponding 9 condensed Jaccard's coefficients
}
\description{
Calculates the 9 Jaccard's Delta coefficients from an array of genotypes
generated with \code{\link{drop.along.ped}}
}
\details{
given two individuals, X with alleles a and b,
and Y with alleles c and d,

delta1-> a==b==c==d
delta2-> a==b, c==d, a!=c
delta3-> a==b==c, a!=d
delta4-> a==b, a=!c, a!=d, c!=d
delta5-> a==c==d, a!=b
delta6-> c==d, c!=a, c!=b, a!=b
delta7-> a==c, b==d, a!=b
delta8-> a==c, a!=b,d, b!=d
delta9-> a!=b, a!=c, a!=d, b!=c, b!=d, c!=d

The sum of all 9 coefficients for any pair of individuals is 1

coancestry=delta1+(delta3+delta5+delta7)/2+delta8/4


ks (kappas) defined for non inbred individuals only
in which case

k0=delta9
k1=delta8
k2=delta7
}
\examples{
\dontrun{}
}
\references{

}
\author{
Jerome Goudet \email{jerome.goudet@unil.ch}
}
