% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropAlongPed.R
\name{kinship.gold}
\alias{kinship.gold}
\title{Estimates gold kinship}
\usage{
kinship.gold(dat)
}
\arguments{
\item{dat}{an array (niXnlX2) produced by \code{\link{drop.along.ped}} containing
which alleles each individual has inherited from the founders}
}
\value{
'gold' kinship matrix, including self kinship on the diagonal
}
\description{
Estimates gold kinship from the genotypes generated with \code{\link{drop.along.ped}}
assuming the founders are unrelated
}
\details{

}
\examples{
\dontrun{}
}
\references{

}
\author{
Jerome Goudet \email{jerome.goudet@unil.ch}
}
