% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maketraits.R
\name{make.traits}
\alias{make.traits}
\title{Creates a phenotype from randomly chosen SNPs}
\usage{
make.traits(bed,n.causal=1000,h2=0.5,minp=0.01,sel=0.0,pop=NULL,popid=NULL)
}
\arguments{
\item{bed}{a bed object}

\item{n.causal}{number of causal loci}

\item{h2}{intended trait heritability}

\item{minp}{minimum maf of a causal allele}

\item{sel}{if !=0 effect size will be scaled by \eqn{1/(2*p*(1-p))^(sel/2)}}

\item{pop}{variance of population effet, relative to VP}

\item{popid}{if NULL, popid taken from bed@ped$famid}
}
\value{
A list with components

\itemize{
\item h2 he intended heritability
\item h2hat the realized heritability
\item trait a data frame of two columns,
gval being the individual genetic values and
pheno their phenotypes
\item causal a data frame with three columns,
chr the chromosome where the causal loci are located;
id the SNP id; and efs their effect size

}
}
\description{
Generates a phenotype from a random subset of SNPs with the required heritability.
possibly including
a (random) population specific effect.
}
\details{
Traits are built using additive effects only. Genetic
values are the sum of the products of dosages at causal loci by their effect sizes.

Additive variance VA is measured as the variance of genetic values. A normal
deviate with variance \eqn{VA * (1/h^2-1)}, where  \eqn{h^2} is the intended trait
heritability, is  added to the genetic values to make the phenotypes. If pop!=NULL,
a random deviate with variance proportional to the phenotypic variance VP (i.e.
\eqn{Vpop=pop*VP}) is added to the individual phenotypes in each population.


If sel is !=0, effect size will be scaled by  \eqn{1/(2*p*(1-p))^(sel/2)}, where p is the
sample frequency of the derived allele.
pop is the variance of the population effect, relative to VP.
if popid is NULL, popid taken from bed@ped$famid
}
