% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thet.bet.2pops.R
\name{thet.bet.2pops}
\alias{thet.bet.2pops}
\title{Expected coancestries and betas from a two population model}
\usage{
thet.bet.2pops(mu=10^-8,m1=10^-4,m2=10^-4,n1=1000,n2=1000,ngen=100,plotit=TRUE)
}
\arguments{
\item{mu}{mutation rate}

\item{m1}{immigration into pop1}

\item{m2}{immigration into pop2}

\item{n1}{size of pop1}

\item{n2}{size of pop2}

\item{ngen}{number of generations}

\item{plotit}{whether to plot the resulting object}
}
\value{
a list with elements
\itemize{
\item Th a matrix of coancestries \eqn{\theta} for pop1, pop2 and pop1-2 (columns) and generations (rows)
\item Be a matrix of \eqn{\beta}s for pop1, pop2 and average (columns) and generations (rows)
}
}
\description{
Quantifies expected coancestries and betas (coancestries relative to
coancestries between population, see
\href{http://www.genetics.org/content/206/4/2085}{Weir and Goudet
(2017) Genetics}) in a two populations model.
}
\examples{

\dontrun{

 # reproduces fig 1 in Weir and Goudet (2017)
 #top row
 thet.bet.2pops(mu=0.0,m2=0.0,m1=0.0,n1=10000,n2=100,ngen=10000)
 #middle row
 thet.bet.2pops(mu=0.001,m2=0.0,m1=0.0,n1=10000,n2=100,ngen=10000)
 #bottom row
 thet.bet.2pops(mu=0.001,m2=0.0,m1=0.01,n1=10000,n2=100,ngen=10000)

   }


}
\references{
\href{http://www.genetics.org/content/206/4/2085}{Weir and Goudet
(2017) Genetics 206:2085}) A Unified Characterization
of Population Structure and Relatedness.
}
