% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibd2dos.R
\name{ibd2dos}
\alias{ibd2dos}
\title{convert Identity By Descent gametes to dosage data}
\usage{
ibd2dos(gametes,n.founders,g.founders=NULL,b1=1,b2=10)
}
\arguments{
\item{gametes}{a nixnlx2 array of gametes created by \code{\link{drop.along.ped}}}

\item{n.founders}{the number of founders in the pedigree}

\item{g.founders}{a (2*nfounders)xnl array with the bi-allelic gametes of the founders (default to NULL)}

\item{b1}{the 1st parameter of the \code{\link{beta}} distribution}

\item{b2}{the 2nd parameter of the \code{\link{beta}} distribution}
}
\value{
a dosage matrix with ni rows and nl columns
}
\description{
Given gametes, the output of \code{\link{drop.along.ped}} on a pedigree and
and n.founders, the number of founders in the pedigree,
return a matrix ni x nl of dosages after having mapped
the 2n.founders alleles at each locus to 0 or 1.
The gametes of the founders are given in g.founders or
frequencies at each locus are drawn from a \code{\link{beta}}
distribution with parameters b1 and b2.
}
\details{
The gametes of  the founders can be entered as a (2*nfounders) x nl matrix. This insures
any haplotype structure in the founders can be kept.  If g.founders is NULL, allelic frequencies
in the founders are drawn from a beta distribution with parameters b1 (default=1)
and b2 (default=10).
}
\examples{
\dontrun{
ped1<-buildped.2sexes(founders.m=10,founders.f=10,fert=1,death.rate=0.5,breed.prop=0.2,n.tstep=5)
ibd.ped1<-drop.along.ped(ped1,nloc=5000,maplength=20)
dos.ped1<-ibd2dos(ibd.ped1,n.founders=20)
}

}
\author{
Jerome Goudet \email{jerome.goudet@unil.ch}
}
