% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropAlongPed.R
\name{jaccard}
\alias{jaccard}
\title{Calculates the 9 Jaccard's Delta coefficients}
\usage{
jaccard(dat)
}
\arguments{
\item{dat}{an nixnlx2 array of genotypes ( ni individuals, nl loci) generated with \code{\link{drop.along.ped}}}
}
\value{
a list of 9 matrices, the lower half of which contain
the corresponding 9 condensed Jaccard's coefficients
}
\description{
Calculates the 9 Jaccard's Delta coefficients from an array of genotypes
generated with \code{\link{drop.along.ped}}
}
\details{
given two individuals, X with alleles a and b,
and Y with alleles c and d,
\itemize{
\item{\eqn{\Delta_1}}{ a==b==c==d}
\item{\eqn{\Delta_2}}{ a==b, c==d, a!=c}
\item{\eqn{\Delta_3}}{ a==b==c, a!=d}
\item{\eqn{\Delta_4}}{ a==b, a=!c, a!=d, c!=d}
\item{\eqn{\Delta_5}}{ a==c==d, a!=b}
\item{\eqn{\Delta_6}}{ c==d, c!=a, c!=b, a!=b}
\item{\eqn{\Delta_7}}{ a==c, b==d, a!=b}
\item{\eqn{\Delta_8}}{ a==c, a!=b,d, b!=d}
\item{\eqn{\Delta_9}}{ a!=b, a!=c, a!=d, b!=c, b!=d, c!=d}
}
The sum of all 9 coefficients for any pair of individuals is 1

Coancestry \eqn{\theta} is
\deqn{\theta=\Delta_1+(\Delta_3+\Delta_5+\Delta_7)/2
+\Delta_8/4}


ks (kappas) defined for non inbred individuals only
in which case
\itemize{
\item{\eqn{k_0}}{ \eqn{\Delta_9}}
\item{\eqn{k_1}}{ \eqn{\Delta_8}}
\item{\eqn{k_2}}{ \eqn{\Delta_7}}
}
}
\examples{
\dontrun{}
}
\references{

}
\author{
Jerome Goudet \email{jerome.goudet@unil.ch}
}
