% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kc0.R
\name{Kc0}
\alias{Kc0}
\title{Estimates the standard kinship matrix}
\usage{
Kc0(dos,matching=FALSE)
}
\arguments{
\item{dos}{either a dosage matrix or a matching matrix}

\item{matching}{TRUE if dos is a matching matrix (nixni),
FALSE (default) if it is a dosage (nixnl) matrix}
}
\value{
the standard, ratio of averages, kinship matrix
}
\description{
Estimates the standard kinship Kc0 matrix as
a ratio of averages
}
\seealso{
\code{\link[hierfstat]{matching}} for allele sharing,
\code{\link[hierfstat]{beta.dosage}} for allele sharing kinships;
\code{\link[gaston]{GRM}} for standard kinship as an average of ratio (default)
}
