% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kc0.R
\name{Kc0}
\alias{Kc0}
\title{Estimates the standard kinship matrix}
\usage{
KC0(dos,inb=FALSE,matching=FALSE)
}
\arguments{
\item{dos}{either a dosage matrix or a matching matrix}

\item{inb}{whether to report inbreeding coefficient
or self coancestries on the main diagonal}

\item{matching}{TRUE if dos is a matching matrix (nixni),
FALSE (default) if it is a dosage (nixnl) matrix}
}
\value{
the standard, ratio of averages, kinship matrix
}
\description{
Estimates the standard kinship Kc0 matrix as
a ratio of averages
}
\seealso{

}
