% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropAlongPed.R
\name{kinship.gold}
\alias{kinship.gold}
\title{Estimates realized (gold) kinship}
\usage{
kinship.gold(dat)
}
\arguments{
\item{dat}{an array (niXnlX2) produced by \code{\link{drop.along.ped}} containing
which alleles each individual has inherited from the founders}
}
\value{
'gold' kinship matrix,  including self kinship on the diagonal
}
\description{
Estimates realized kinship from the genotypes generated with \code{\link{drop.along.ped}}
assuming the founders are unrelated
}
\details{
'gold' kinship is the realized proportion of shared alleles between individuals,
using the founders as reference (each founder is assigned two unique alleles at each position).
the realized inbreeding coefficient of an individual \eqn{F_i} is obtained from its self-kinship \eqn{k_{ii}} as
\eqn{F_i=2 k_{ii}-1}
}
\examples{
\dontrun{}
}
\references{

}
\author{
Jerome Goudet \email{jerome.goudet@unil.ch}
}
