% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedARM.R
\name{pedARM}
\alias{pedARM}
\title{calculates the additive relationship matrix from a pedigree}
\usage{
pedARM(s,d,fcoan=NULL)
}
\arguments{
\item{s}{a vector with sire identies}

\item{d}{a vector wit dam identities}

\item{fcoan}{a matrix of coancestries for the founders, if available}
}
\value{
a matrix with the expected relatedness
among all airs of individuals and twice self coancestry
on the diagonal
}
\description{
given a vector of sires and dams from a pedigree, calculates
the Additive relationship matrix
}
